+++ src/llm-providers/openai/OpenAiAssistantExecutionToolsOptions.ts
@@ -1,0 +1,17 @@
+import type { ClientOptions } from 'openai';
+import type { string_token } from '../../types/typeAliases';
+import { OpenAiExecutionToolsOptions } from './OpenAiExecutionToolsOptions';
+
+/**
+ * Options for `OpenAiAssistantExecutionTools`
+ *
+ * @public exported from `@promptbook/openai`
+ */
+export type OpenAiAssistantExecutionToolsOptions = OpenAiExecutionToolsOptions &
+    ClientOptions & {
+        /**
+         * Which assistant to use
+         */
+        assistantId?: string_token;
+        // <- TODO: [🧠] This should be maybe more like model for each prompt?
+    };