+++ src/llm-providers/openai/computeOpenAiUsage.ts
@@ -25,6 +25,7 @@ export function computeOpenAiUsage(
             | OpenAI.Chat.Completions.ChatCompletion
             | OpenAI.Completions.Completion
             | OpenAI.Embeddings.CreateEmbeddingResponse,
+            // !!!!!! | OpenAI.Beta.Threads.Messages.Message,
             'model' | 'usage'
         >
     >,